add voices

+++ main.py
@@ -40,29 +40,39 @@ async def choose_dir(message):
 @dp.callback_query_handler(dir_set.filter())
 async def choose_voice(call: types.CallbackQuery, callback_data: dict):
     user_id = callback_data['user_id']
-    dir_name = callback_data['dir']
-    db[user_id] = {'dir': dir_name}
 
-    path = 'voices/' + dir_name
-    voices = sorted(os.listdir(path))
+    if call.from_user.id != int(user_id):
+        await call.answer('Куда лезешь мудила')
+        await bot.answer_callback_query(call.id)
+    else:
+        dir_name = callback_data['dir']
+        db[user_id] = {'dir': dir_name}
 
-    keyboard = InlineKeyboardMarkup()
-    for item in voices:
-        item_cut = item[:-4]
-        keyboard.add(InlineKeyboardButton(item_cut, callback_data=voice_set.new(user_id=user_id, voice=item)))
+        path = 'voices/' + dir_name
+        voices = sorted(os.listdir(path))
 
-    await call.message.edit_text(dir_name, reply_markup=keyboard)
+        keyboard = InlineKeyboardMarkup()
+        for item in voices:
+            item_cut = item[:-4]
+            keyboard.add(InlineKeyboardButton(item_cut, callback_data=voice_set.new(user_id=user_id, voice=item)))
+
+        await call.message.edit_text(dir_name, reply_markup=keyboard)
 
 
 @dp.callback_query_handler(voice_set.filter())
 async def send_voice(call: types.CallbackQuery, callback_data: dict):
     user_id = callback_data['user_id']
-    dir_name = db[user_id]['dir']
-    voice = callback_data['voice']
-    path = f'voices/{dir_name}/{voice}'
 
-    await call.message.delete()
-    await bot.send_voice(call.message.chat.id, InputFile(path))
+    if call.from_user.id != int(user_id):
+        await call.answer('Куда лезешь мудила', show_alert=True)
+        await bot.answer_callback_query(call.id)
+    else:
+        dir_name = db[user_id]['dir']
+        voice = callback_data['voice']
+        path = f'voices/{dir_name}/{voice}'
+
+        await call.message.delete()
+        await bot.send_voice(call.message.chat.id, InputFile(path))
 
 
 async def on_startup(_):
